services/audience_segmentation_service/requirements.txt
fastapi>=0.100.0,<0.112.0
uvicorn[standard]>=0.20.0,<0.30.0
pydantic>=2.0.0,<3.0.0

# ML Pipeline Dependencies (as per your plan)
sentence-transformers>=2.2.0,<3.0.0 # For Sentence-BERT models like all-MiniLM-L6-v2
umap-learn>=0.5.0,<0.6.0
scikit-learn>=1.0.0,<1.6.0 # For K-Means++ and other utilities
transformers>=4.0.0,<5.0.0 # For Hugging Face models like BART for zero-shot classification
numpy>=1.20.0,<1.27.0

# Optional: for n-gram extraction if not using scikit-learn's
# nltk>=3.6.0,<4.0.0

# Consider adding specific versions later once compatibility is tested 